New features for upcoming 1.2

+++ luna/osimage.py
@@ -20,21 +20,20 @@ along with Luna.  If not, see <http://www.gnu.org/licenses/>.
 
 '''
 
-from config import *
+from config import torrent_key
 
 import os
 import sys
 import pwd
-import grp
 import rpm
 import uuid
 import shutil
 import logging
 import tempfile
+import traceback
 import subprocess
 import libtorrent
-
-from bson.dbref import DBRef
+import exceptions
 
 from luna.base import Base
 from luna.cluster import Cluster
@@ -65,7 +64,7 @@ class OsImage(Base):
                          'dracutmodules': type(''), 'tarball': type(''),
                          'torrent': type(''), 'kernfile': type(''),
                          'initrdfile': type(''), 'grab_exclude_list': type(''),
-                         'grab_filesystems': type('')}
+                         'grab_filesystems': type(''), 'comment': type('')}
 
         # Check if this osimage is already present in the datastore
         # Read it if that is the case
@@ -116,7 +115,7 @@ class OsImage(Base):
                        'dracutmodules': 'luna,-i18n,-plymouth',
                        'kernmodules': 'ipmi_devintf,ipmi_si,ipmi_msghandler',
                        'grab_exclude_list': grab_list_content,
-                       'grab_filesystems': '/,/boot'}
+                       'grab_filesystems': '/,/boot', 'comment': None}
 
             self.log.debug("Saving osimage '{}' to the datastore"
                            .format(osimage))
@@ -171,16 +170,20 @@ class OsImage(Base):
 
         try:
             # dirty, but 4 times faster
-            tar_out = subprocess.Popen(['/usr/bin/tar',
-                                        '-C', '/',
-                                        '--one-file-system',
-                                        '--xattrs',
-                                        '--selinux',
-                                        '--acls',
-                                        '--checkpoint=100',
-                                        '--exclude=./tmp/' + tarfile,
-                                        '-c', '-z', '-f', '/tmp/' + tarfile, '.'],
-                                       stderr=subprocess.PIPE)
+            tar_out = subprocess.Popen(
+                [
+                    '/usr/bin/tar',
+                    '-C', '/',
+                    '--one-file-system',
+                    '--xattrs',
+                    '--selinux',
+                    '--acls',
+                    '--checkpoint=100',
+                    '--exclude=./tmp/' + tarfile,
+                    '-c', '-z', '-f', '/tmp/' + tarfile, '.'
+                ],
+                stderr=subprocess.PIPE
+            )
 
             stat_symb = ['\\', '|', '/', '-']
             i = 0
@@ -193,14 +196,23 @@ class OsImage(Base):
                 sys.stdout.write('\r')
 
         except:
-            os.remove('/tmp/' + tarfile)
+            exc_type, exc_value, exc_traceback = sys.exc_info()
+            if exc_type == exceptions.KeyboardInterrupt:
+                self.log.error('Keyboard interrupt.')
+            else:
+                self.log.error(exc_value)
+                self.log.debug(traceback.format_exc())
+
+            if os.path.isfile('/tmp/' + tarfile):
+                os.remove('/tmp/' + tarfile)
+
             sys.stdout.write('\r')
 
             os.fchdir(real_root)
             os.chroot(".")
             os.close(real_root)
 
-            return None
+            return False
 
         os.fchdir(real_root)
         os.chroot(".")
@@ -326,11 +338,14 @@ class OsImage(Base):
         real_root = os.open("/", os.O_RDONLY)
         os.chroot(image_path)
 
+        dracut_succeed = True
+
         try:
             dracut_modules = subprocess.Popen(['/usr/sbin/dracut', '--kver',
                                                kernver, '--list-modules'],
                                               stdout=subprocess.PIPE)
             luna_exists = False
+
             while dracut_modules.poll() is None:
                 line = dracut_modules.stdout.readline()
                 if line.strip() == 'luna':
@@ -350,21 +365,33 @@ class OsImage(Base):
                 line = create.stdout.readline()
 
         except:
-            self.log.error("Error while building initrd.")
-            os.fchdir(real_root)
-            os.chroot(".")
-            os.close(real_root)
-            cleanup_mounts(image_path)
-            return False
+            dracut_succeed = False
+
+        if create.returncode:
+            dracut_succeed = False
 
         os.fchdir(real_root)
         os.chroot(".")
         os.close(real_root)
         cleanup_mounts(image_path)
 
-        shutil.copy(image_path + tmp_path + '/' + initrdfile, path_to_store)
-        shutil.copy(image_path + '/boot/vmlinuz-' + kernver,
-                    path_to_store + '/' + kernfile)
+        if not dracut_succeed:
+            self.log.error("Error while building initrd.")
+            return False
+
+        initrd_path = image_path + tmp_path + '/' + initrdfile
+        kernel_path = image_path + '/boot/vmlinuz-' + kernver
+
+        if not os.path.isfile(kernel_path):
+            self.log.error("Unable to find kernel in {}".format(kernel_path))
+            return False
+
+        if not os.path.isfile(initrd_path):
+            self.log.error("Unable to find initrd in {}".format(initrd_path))
+            return False
+
+        shutil.move(initrd_path, path_to_store + '/' + initrdfile)
+        shutil.copy(kernel_path, path_to_store + '/' + kernfile)
         os.chown(path_to_store + '/' + initrdfile, user_id, grp_id)
         os.chmod(path_to_store + '/' + initrdfile, 0644)
         os.chown(path_to_store + '/' + kernfile, user_id, grp_id)
@@ -376,7 +403,6 @@ class OsImage(Base):
         return True
 
     def copy_boot(self):
-        tmp_path = '/tmp'  # in chroot env
         image_path = self.get('path')
         kernver = self.get('kernver')
         initrdfile = self.name + '-initramfs-' + kernver
@@ -391,14 +417,24 @@ class OsImage(Base):
         path = cluster.get('path')
         path_to_store = path + "/boot"
 
+        initrd_path = image_path + '/boot/initramfs-' + kernver + '.img'
+        kernel_path = image_path + '/boot/vmlinuz-' + kernver
+
+        if not os.path.isfile(kernel_path):
+            self.log.error("Unable to find kernel in {}".format(kernel_path))
+            return False
+
+        if not os.path.isfile(initrd_path):
+            self.log.error("Unable to find initrd in {}".format(initrd_path))
+            return False
+
         if not os.path.exists(path_to_store):
             os.makedirs(path_to_store)
             os.chown(path_to_store, user_id, grp_id)
 
-        shutil.copy(image_path + '/boot/initramfs-' + kernver + '.img',
-                    path_to_store + '/' + initrdfile)
-        shutil.copy(image_path + '/boot/vmlinuz-' + kernver,
-                    path_to_store + '/' + kernfile)
+        shutil.copy(initrd_path, path_to_store + '/' + initrdfile)
+        shutil.copy(kernel_path, path_to_store + '/' + kernfile)
+
         os.chown(path_to_store + '/' + initrdfile, user_id, grp_id)
         os.chmod(path_to_store + '/' + initrdfile, 0644)
         os.chown(path_to_store + '/' + kernfile, user_id, grp_id)
